// src/crosscommon/entities/Task.d.ts
import { iEntity } from "../iEntity";
import { FieldDefinition } from "../FieldDefinition";
import { ViewJoinDefinition } from "../ViewJoinDefinition";
export declare class Task implements iEntity {
    tsk_id: string;
    tsk_id_container: string;
    tsk_id_record: string;
    tsk_name: string;
    tsk_notes: string;
    tsk_parent: string;
    tsk_order: number;
    tsk_date_done: Date;
    tsk_total_time_spent: number;
    tsk_ctg_in_process: number;
    tsk_qualifiers: string;
    tsk_tags: string;
    tsk_estimated_duration: number;
    tsk_schedule_date_start: Date;
    tsk_schedule_date_end: Date;
    tsk_date_view_until: Date;
    tsk_id_user_added: string;
    tsk_id_user_asigned: string;
    tsk_template: string;
    tsk_template_state: string;
    tsk_date_due: Date;
    tsk_id_related: string;
    tsk_url: string;
    tsk_ctg_repeats: number;
    tsk_id_main: string;
    tsk_ctg_rep_type: number;
    tsk_ctg_rep_after_completion: number;
    tsk_ctg_rep_end: number;
    tsk_rep_date_end: Date;
    tsk_rep_end_iteration: number;
    tsk_rep_iteration: number;
    tsk_rep_frequency: number;
    tsk_ctg_rep_frequency_rule: number;
    tsk_rep_weekdays: string;
    tsk_date_add: Date;
    tsk_date_mod: Date;
    tsk_ctg_status: number;
    tsk_txt_in_process: string;
    tsk_txt_repeats: string;
    tsk_txt_rep_type: string;
    tsk_txt_rep_after_completion: string;
    tsk_txt_rep_end: string;
    tsk_txt_rep_frequency_rule: string;
    tsk_txt_status: string;
    metadata: {
        name: string;
        namespace: string;
        removeMeans: string;
        authNeeded: boolean;
        displayOnMenu: boolean;
        prefix: string;
        permissionsTemplate: string;
        tableName: string;
        viewName: string;
        permissions: string[];
        specialFeatures: string[];
        fields: FieldDefinition[];
        view: ViewJoinDefinition[];
    };
    constructor(base?: any);
    recordName: () => string;
}
//# sourceMappingURL=Task.d.ts.map